+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Handle multi-level options in the Select
-- Feature: Select defaults to un-clearable, un-searchable, and un-multi
+- Feature: Select can now be clearable, searchable, and multi
 
 ## [v.4.4.0] - 2016-12-27